Make ImportPmx undo compatible
This helps undoing operations after the import. Because currently undoing the operation after the import undos the import as well. This small change creates an undo point after the import is finished.
You might want to add 'REGISTER' and 'UNDO' to every bl_options in order to make every command undoable.